cloud_deployment/requirements.txt
# Dependencias para Google Cloud Functions
functions-framework==3.4.0
google-cloud-bigquery==3.38.0
pandas==2.3.2
pyarrow==21.0.0
requests==2.32.5
numpy==2.3.3
